+++ src/master/flags.hpp
@@ -92,6 +92,7 @@ public:
   std::string authorizers;
   std::string http_authenticators;
   Option<std::string> http_framework_authenticators;
+  Option<std::set<std::string>> http_cors_allowed_origins;
   size_t max_operator_event_stream_subscribers;
   size_t max_completed_frameworks;
   size_t max_completed_tasks_per_framework;